moved stuff from table.js to index.js
