+++ components/Table.js
@@ -6,18 +6,18 @@ import Button from "react-bootstrap/Button";
 
 export const getServerSideProps = optionalAuth;
 
-function Table(props) {
+function Table(props, mealMatrix) {
   const user = props.user;
-  const mealMatrix = [
-    ["", "", ""],
-    ["", "", ""],
-    ["", "", ""],
-    ["", "", ""],
-    ["", "", ""],
-    ["", "", ""],
-    ["", "", ""],
-  ];
-  const { data } = useSWR("/api/meal");
+  // const mealMatrix = [
+  //   ["", "", ""],
+  //   ["", "", ""],
+  //   ["", "", ""],
+  //   ["", "", ""],
+  //   ["", "", ""],
+  //   ["", "", ""],
+  //   ["", "", ""],
+  // ];
+  // const { data } = useSWR("/api/meal");
 
   const removeRecipe = async (e) => {
     e.preventDefault();
@@ -34,25 +34,25 @@ function Table(props) {
     location.reload();
   };
 
-  if (data) {
-    for (let i = 0; i < 21; i++) {
-      if (data[i]) {
-        let dayy = -1;
-        let typee = -1;
-        if (data[i].day == "mon") dayy = 0;
-        if (data[i].day == "tue") dayy = 1;
-        if (data[i].day == "wed") dayy = 2;
-        if (data[i].day == "thu") dayy = 3;
-        if (data[i].day == "fri") dayy = 4;
-        if (data[i].day == "sat") dayy = 5;
-        if (data[i].day == "sun") dayy = 6;
-        if (data[i].type == "break") typee = 0;
-        if (data[i].type == "lunch") typee = 1;
-        if (data[i].type == "dinnr") typee = 2;
-        if (dayy != -1 && typee != -1) mealMatrix[dayy][typee] = data[i];
-      }
-    }
-  }
+  // if (data) {
+  //   for (let i = 0; i < 21; i++) {
+  //     if (data[i]) {
+  //       let dayy = -1;
+  //       let typee = -1;
+  //       if (data[i].day == "mon") dayy = 0;
+  //       if (data[i].day == "tue") dayy = 1;
+  //       if (data[i].day == "wed") dayy = 2;
+  //       if (data[i].day == "thu") dayy = 3;
+  //       if (data[i].day == "fri") dayy = 4;
+  //       if (data[i].day == "sat") dayy = 5;
+  //       if (data[i].day == "sun") dayy = 6;
+  //       if (data[i].type == "break") typee = 0;
+  //       if (data[i].type == "lunch") typee = 1;
+  //       if (data[i].type == "dinnr") typee = 2;
+  //       if (dayy != -1 && typee != -1) mealMatrix[dayy][typee] = data[i];
+  //     }
+  //   }
+  // }
   return (
     <Container user={user}>
       {user ? (
