+++ stories/EnterMeal.stories.js
@@ -1,18 +1,10 @@
 import React from "react";
 import { select, text } from "@storybook/addon-knobs";
 import EnterMeal from "../components/EnterMeal";
-import AppNavbar from "../components/AppNavbar";
 
 export default {
   title: "EnterMeal",
   component: EnterMeal,
-};
-
-export const loggedOutEmpty = () => {
-  return <AppNavbar />;
-  // this ^ is not right. this line is supposed to be
-  // return <EnterMeal />;
-  // but it's having an issue with "nickname" in EnterMeal.js
 };
 
 export const loggedInDisplaysForm = () => {
